move more tests into the snuba-less test file
